Make the Lily-restart feature in LilyLauncher work by two changes:  * enable proper closing in AvroServer (experience will tell if the reason it was previously disabled is still valid)  * make it possible to disable the deletion of HBase connections on shutdown (via a system property)
git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@4938 b80d95f8-e103-0410-bd55-e03a9dc29394

+++ cr/process/server/general-module/src/main/java/org/lilyproject/server/modules/general/HBaseConnectionDisposer.java
@@ -23,10 +23,13 @@ import javax.annotation.PreDestroy;
 public class HBaseConnectionDisposer {
     @PreDestroy
     public void stop() {
-        try {
-            HConnectionManager.deleteAllConnections(true);
-        } catch (Throwable t) {
-            LogFactory.getLog(getClass()).error("Problem cleaning up HBase connections", t);
+        String opt = System.getProperty("lily.hbase.deleteConnections");
+        if (opt == null || opt.equals("true")) {
+            try {
+                HConnectionManager.deleteAllConnections(true);
+            } catch (Throwable t) {
+                LogFactory.getLog(getClass()).error("Problem cleaning up HBase connections", t);
+            }
         }
     }
 }
